fixed: zhCN, zhTW
