--- src/i18n/languages/zh-CN.ts
@@ -1,40 +1,0 @@
-import Key from "../i18nKey";
-import type { Translation } from "../translation";
-
-export const zhCN: Translation = {
-	[Key.home]: "主页",
-	[Key.about]: "关于",
-	[Key.archive]: "归档",
-	[Key.search]: "搜索",
-
-	[Key.tags]: "标签",
-	[Key.categories]: "分类",
-	[Key.recentPosts]: "最新文章",
-
-	[Key.comments]: "评论",
-
-	[Key.untitled]: "无标题",
-	[Key.uncategorized]: "未分类",
-	[Key.noTags]: "无标签",
-
-	[Key.wordCount]: "字",
-	[Key.wordsCount]: "字",
-	[Key.minuteCount]: "分钟",
-	[Key.minutesCount]: "分钟",
-	[Key.postCount]: "篇文章",
-	[Key.postsCount]: "篇文章",
-
-	[Key.themeColor]: "主题色",
-
-	[Key.lightMode]: "亮色",
-	[Key.darkMode]: "暗色",
-	[Key.systemMode]: "跟随系统",
-
-	[Key.more]: "更多",
-
-	[Key.author]: "作者",
-	[Key.publishedAt]: "发布于",
-	[Key.license]: "许可协议",
-	[Key.unsupportedLang]: "该页面尚未翻译成%s。",
-	[Key.langName]: "中文",
-};
