--- src/i18n/languages/zh-TW.ts
@@ -1,40 +1,0 @@
-import Key from "../i18nKey";
-import type { Translation } from "../translation";
-
-export const zhTW: Translation = {
-	[Key.home]: "首頁",
-	[Key.about]: "關於",
-	[Key.archive]: "彙整",
-	[Key.search]: "搜尋",
-
-	[Key.tags]: "標籤",
-	[Key.categories]: "分類",
-	[Key.recentPosts]: "最新文章",
-
-	[Key.comments]: "評論",
-
-	[Key.untitled]: "無標題",
-	[Key.uncategorized]: "未分類",
-	[Key.noTags]: "無標籤",
-
-	[Key.wordCount]: "字",
-	[Key.wordsCount]: "字",
-	[Key.minuteCount]: "分鐘",
-	[Key.minutesCount]: "分鐘",
-	[Key.postCount]: "篇文章",
-	[Key.postsCount]: "篇文章",
-
-	[Key.themeColor]: "主題色",
-
-	[Key.lightMode]: "亮色",
-	[Key.darkMode]: "暗色",
-	[Key.systemMode]: "跟隨系統",
-
-	[Key.more]: "更多",
-
-	[Key.author]: "作者",
-	[Key.publishedAt]: "發佈於",
-	[Key.license]: "許可協議",
-	[Key.unsupportedLang]: "該頁面尚未翻譯成%s。",
-	[Key.langName]: "繁體中文",
-};
