+++ src/i18n/languages/zh_CN.ts
@@ -1,7 +1,7 @@
 import Key from "../i18nKey";
 import type { Translation } from "../translation";
 
-export const zh_CN: Translation = {
+export const zhCN: Translation = {
 	[Key.home]: "主页",
 	[Key.about]: "关于",
 	[Key.archive]: "归档",
@@ -35,4 +35,6 @@ export const zh_CN: Translation = {
 	[Key.author]: "作者",
 	[Key.publishedAt]: "发布于",
 	[Key.license]: "许可协议",
+	[Key.unsupportedLang]: "该页面尚未翻译成%s。",
+	[Key.langName]: "中文",
 };
